clock: Do not use timezone local key to register its change
SYSTEM_SETTINGS_KEY_LOCAL_TIMEZONE is a key only used as getter and
can not be used to register callback on its change. However when
the timezone is changed, system setting calls event registered on
SYSTEM_SETTINGS_KEY_TIME_CHANGED so removing this key does not
effect on shown hour.

Change-Id: Idb7102b5029d9dc71275d528124876fd1fcf1715